pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "acd"
version = "0.1.0"
description = "Analysis-ready data workflows for Australian Climate Data - NetCDF to Zarr conversion pipeline"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Thomas Moore", email = "thomas@example.com"}
]
requires-python = ">=3.11"
dependencies = [
    "xarray",
    "zarr",
    "dask",
    "dask-jobqueue",
    "rechunker",
    "papermill",
    "netcdf4",
    "pyyaml",
    "click",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "ruff",
]

[project.scripts]
acd-manifest = "acd.cli.manifest:main"
acd-base = "acd.cli.base:main"
acd-rechunk = "acd.cli.rechunk:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = []
